% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majestic.R
\name{majestic}
\alias{majestic}
\title{Majestic -
This function provides all backlink data of an URL, including its CitationFlow, TrustFlow, and Root domain.}
\usage{
majestic(url, api_key)
}
\arguments{
\item{url}{An URL you'd like to investigate.}

\item{api_key}{An identification key assigned to a user after subscribing to Majestic that is available via Profile page.}
}
\value{
majestic provides all backlink data of an URL
}
\description{
Majestic -
This function provides all backlink data of an URL, including its CitationFlow, TrustFlow, and Root domain.
}
\examples{
\dontrun{
majestic("https://dataseolabs.com/fr/","xxxxxxxxxxxxxxxxxxx")
}
}
