% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majestic_domain.R
\name{majestic_domain}
\alias{majestic_domain}
\title{Majestic Domain}
\usage{
majestic_domain(url, api_key)
}
\arguments{
\item{url}{An URL you'd like to investigate the domain data.}

\item{api_key}{An identification key assigned to a user after subscribing to Majestic that is available via Profile page.}
}
\value{
majestic_domain provides all backlink data of an URL
}
\description{
This function first extract the root domain from an URL, then provides all backlink data of the domain, including its CitationFlow, TrustFlow, and Root domain.
}
\examples{
\dontrun{
url <- "https://dataseolabs.com/en/"
api_key <- "xxxxxxxxxxxxxxxxxxx"
majestic_domain(url, api_key)
}
}
