% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majestic_url.R
\name{majestic_url}
\alias{majestic_url}
\title{Majestic URL -}
\usage{
majestic_url(url, api_key)
}
\arguments{
\item{url}{An URL you'd like to investigate.}

\item{api_key}{An identification key assigned to a user after subscribing to Majestic that is available via Profile page.}
}
\value{
majestic_url provides all backlink data of an URL
}
\description{
This function provides all backlink data of an URL, including its CitationFlow, TrustFlow, and Root domain.
}
\examples{
\dontrun{
url <- "https://dataseolabs.com/en/"
api_key <- "xxxxxxxxxxxxxxxxxxx"
majestic_domain(url, api_key)
}
}
